fix(drawback): Simplify nightly discussion
In part, this reverts 545ef5257edcfb0d41c62bb8498c4fac6b4087d2.

I don't want to get into an argument over the right way we should be
discouraging nightly shenanigans or how much of a use case there is or
not.
That is all for future possibilities.
The main point is that we aren't covering a use case from one piece of
prior art.