Allow no-effect unnamed variants as a lint, add alternative
The `taken_discrimimant_ranges` and `empty_discriminant_ranges` lints
catch the situation in which an unnamed variant has no effect. The
latter is much more dangerous than the former, so it is
`deny`-by-default. However, both should be allowed for certain codegen
and macro cases as described.

This also moves around and expands some language regarding `derive`
difficulty in open-enum, as well as add a suggested alternative.